added the start of a script to generate the files from the Java Edition data
If 'from_jar' is defined in the system arguments this script will look for the server.jar file and a version of java to export the vanilla files.
It will then find all the relevant data and reform it into the needed form and export to individual json files.
In the future this script will also load from custom json files to modify the output data