[Compiler] Added memory buffer and async write
The compiler was starting to get a bit slow with the amount of files it was having to read and write.

Added a memory buffer which is written to and saved out at the end rather than constantly reading and writing from disk.
Using threads to write in parrallel. This seems to speed things up a bit