Fixed slabs in Bedrock 1.19
This was a more complex fix than I initially thought.
The block states changed but the block version did not.
This means that when loading the newest translator with that block version is used (eg 1.18 will use 1.19)
When converting back the translator is looked up by version tuple.
This meant that the 1.19 translator needed to be able to load the 1.18 and 1.19 states for slabs.