Backport the Bedrock block alias names
There have been some cases of the new block alias names appearing under the old version.
This backports the new names back to 1.16.
If the new names are found in the old version they will be converted to the correct universal state.
When saving back they will be saved to the correct state for that version.